Added subroute handling
Moved the route handling to a new Subrouter. This handles every request,
apply the defined middleware and return the appropiate handler.

The middleware can be global and applied to every request or per
Subrouter, that is, will be applied only to those subroutes.

The router/muxer than handle ther routes is fixed: httprouter